plot.py: lack of support for both R and W operations
plot.py was trying to grab configuration information from tepmlate.ini,
but that does not contain information about configurations anymore (it
use to do in the past). Instead, the script is now parsing the 'bs' and
the 'iodepth' and is generating all the possible configurations in order
to read the results from the csv files.

Trivial:
 * Removing deprecated comments
 * Adding the label 'configuration' to the X axis
 * Saving the image to the dislk